# docs/openapi/desktop-v1.2026-08-11.yaml
openapi: 3.1.0
info:
  title: Klui Desktop API
  version: desktop-v1.2026-08-11
  description: Immutable Windows-first contract. OAuth tokens are opaque to native clients.
servers:
  - url: https://klui.tech
paths:
  /oauth/desktop/authorize:
    get:
      operationId: authorizeDesktop
      parameters:
        - { name: client_id, in: query, required: true, schema: { type: string, enum: [klui-desktop-windows, klui-desktop-macos] } }
        - { name: redirect_uri, in: query, required: true, schema: { type: string } }
        - { name: response_type, in: query, required: true, schema: { type: string, const: code } }
        - { name: code_challenge, in: query, required: true, schema: { type: string, minLength: 43, maxLength: 128 } }
        - { name: code_challenge_method, in: query, required: true, schema: { type: string, const: S256 } }
        - { name: state, in: query, required: true, schema: { type: string, minLength: 20, maxLength: 512 } }
      responses:
        "302": { description: Redirect to the configured hosted authorization server }
  /oauth/desktop/token:
    post:
      operationId: exchangeDesktopToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type, client_id]
              properties:
                grant_type: { type: string, enum: [authorization_code, refresh_token] }
                client_id: { type: string, enum: [klui-desktop-windows, klui-desktop-macos] }
                code: { type: string }
                redirect_uri: { type: string }
                code_verifier: { type: string }
                refresh_token: { type: string }
      responses:
        "200":
          description: Normalized rotating token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "400":
          description: OAuth error, including invalid_grant
  /api/desktop/v1/me:
    get:
      operationId: getDesktopAccount
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ClientVersion" }
      responses:
        "200":
          description: Canonical Klui account and capabilities
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "426": { $ref: "#/components/responses/UpdateRequired" }
  /api/desktop/v1/chat/completions:
    post:
      operationId: createDesktopChatCompletion
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/RequestId" }
        - { $ref: "#/components/parameters/ClientVersion" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages, stream]
              properties:
                model: { type: string, const: klui-desktop-agent }
                stream: { type: boolean, const: true }
                messages:
                  type: array
                  minItems: 1
                  maxItems: 128
                  description: Up to four embedded PNG, JPEG, or WebP images; each image is at most 4 MiB and 2560 pixels on its longest edge. Remote image URLs are rejected.
                  items: { type: object }
                tools: { type: array, maxItems: 64, items: { type: object } }
                tool_choice: {}
      responses:
        "200":
          description: OpenAI-compatible server-sent event stream
          content:
            text/event-stream: {}
        "401": { $ref: "#/components/responses/Error" }
        "402": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "426": { $ref: "#/components/responses/UpdateRequired" }
        "429": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/desktop/v1/speech-to-text:
    post:
      operationId: transcribeDesktopSpeech
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/RequestId" }
        - { $ref: "#/components/parameters/ClientVersion" }
      requestBody:
        required: true
        content:
          audio/wav:
            schema: { type: string, contentEncoding: binary, maxLength: 5242880 }
      responses:
        "200":
          description: Transcript and settled duration usage
          content:
            application/json:
              schema:
                type: object
                required: [transcript, usage]
                properties:
                  transcript: { type: string }
                  usage:
                    type: object
                    required: [credits, durationSeconds]
                    properties:
                      credits: { type: number }
                      durationSeconds: { type: number, maximum: 30 }
        "401": { $ref: "#/components/responses/Error" }
        "402": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "426": { $ref: "#/components/responses/UpdateRequired" }
        "429": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
  /api/desktop/v1/logout:
    post:
      operationId: logoutDesktop
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ClientVersion" }
      responses:
        "200":
          description: Current provider session revoked when supported
        "401": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RequestId:
      name: X-Klui-Request-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
    ClientVersion:
      name: X-Klui-Client-Version
      in: header
      required: true
      schema: { type: string, pattern: "^\\d+\\.\\d+\\.\\d+(?:[-+].*)?$" }
  schemas:
    TokenResponse:
      type: object
      required: [access_token, refresh_token, expires_in, token_type]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, minimum: 1 }
        token_type: { type: string, const: Bearer }
        scope: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            retryable: { type: boolean }
    MeResponse:
      type: object
      required: [account, usage, privacy, capabilities, versions]
      properties:
        account:
          type: object
          required: [id, email]
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
        subscription: { type: [object, "null"] }
        plan: { type: [object, "null"] }
        usage:
          type: object
          required: [used, reserved, remaining, limit, limited]
          properties:
            used: { type: number }
            reserved: { type: number }
            remaining: { type: number }
            limit: { type: number }
            limited: { type: boolean }
            resetAt: { type: [string, "null"] }
        privacy:
          type: object
          required: [policyVersion, consentRequired]
          properties:
            policyVersion: { type: string }
            consentRequired: { type: boolean }
        capabilities:
          type: object
          required: [chat, voice, computerUse, maxAgentIterations, maxCompletionTokens]
          properties:
            chat: { type: boolean }
            voice: { type: boolean }
            computerUse: { type: boolean }
            maxAgentIterations: { type: integer, maximum: 20 }
            maxCompletionTokens: { type: integer, maximum: 8192 }
        versions:
          type: object
          required: [minimum, latest, downloadUrl]
          properties:
            minimum: { type: string }
            latest: { type: string }
            downloadUrl: { type: string, format: uri }
  responses:
    Error:
      description: Stable desktop error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UpdateRequired:
      description: Client update is required
      content:
        application/json:
          schema:
            allOf:
              - { $ref: "#/components/schemas/ErrorResponse" }
              - type: object
                required: [minimum_version, latest_version, download_url]
                properties:
                  minimum_version: { type: string }
                  latest_version: { type: string }
                  download_url: { type: string, format: uri }